+++ src/main/java/de/mossgrabers/convertwithmoss/format/akai/s1000/AkaiProgram.java
@@ -2,7 +2,7 @@
 // (c) 2019-2026
 // Licensed under LGPLv3 - http://www.gnu.org/licenses/lgpl-3.0.txt
 
-package de.mossgrabers.convertwithmoss.format.akai.s1000s3000;
+package de.mossgrabers.convertwithmoss.format.akai.s1000;
 
 import java.io.IOException;
 import java.util.List;
@@ -168,7 +168,7 @@ public class AkaiProgram extends AkaiDiskElement
      * @param dirEntry The directory entry of the program
      * @throws IOException Could not read the program
      */
-    public AkaiProgram (final AkaiDiskImage disk, final AkaiVolume parent, final AkaiDirEntry dirEntry) throws IOException
+    public AkaiProgram (final AkaiS1000DiskImage disk, final AkaiVolume parent, final AkaiDirEntry dirEntry) throws IOException
     {
         super (disk.getPos ());