+++ src/main/java/de/mossgrabers/convertwithmoss/format/akai/s1000/AkaiStreamWhence.java
@@ -2,7 +2,7 @@
 // (c) 2019-2026
 // Licensed under LGPLv3 - http://www.gnu.org/licenses/lgpl-3.0.txt
 
-package de.mossgrabers.convertwithmoss.format.akai.s1000s3000;
+package de.mossgrabers.convertwithmoss.format.akai.s1000;
 
 /**
  * Relative markers for positioning the read-position.