+++ src/main/java/de/mossgrabers/convertwithmoss/format/akai/s1000/AkaiVolume.java
@@ -2,7 +2,7 @@
 // (c) 2019-2026
 // Licensed under LGPLv3 - http://www.gnu.org/licenses/lgpl-3.0.txt
 
-package de.mossgrabers.convertwithmoss.format.akai.s1000s3000;
+package de.mossgrabers.convertwithmoss.format.akai.s1000;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -36,7 +36,7 @@ public class AkaiVolume extends AkaiDiskElement
      * @param dirEntry The directory entry of the volume
      * @throws IOException Could not read the volume
      */
-    AkaiVolume (final AkaiDiskImage disk, final AkaiPartition partition, final AkaiDirEntry dirEntry) throws IOException
+    AkaiVolume (final AkaiS1000DiskImage disk, final AkaiPartition partition, final AkaiDirEntry dirEntry) throws IOException
     {
         this.partition = partition;
         this.dirEntry = dirEntry;