+++ src/main/java/de/mossgrabers/convertwithmoss/format/yamaha/ysfc/file/YamahaYsfcChunk.java
@@ -77,6 +77,8 @@ public class YamahaYsfcChunk
     public void read (final InputStream in, final int version) throws IOException
     {
         this.chunkID = StreamUtils.readASCII (in, 4);
+        System.out.println ("\n" + this.chunkID);
+
         this.chunkLength = (int) StreamUtils.readUnsigned32 (in, true);
         this.numItemsInChunk = (int) StreamUtils.readUnsigned32 (in, true);